Moved helpers to `utils::hir_parse` module
