+++ bevy_lint/src/utils/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod hir_parse;
